// src/rpg-core/types/combat-action.ts
export enum CombatAction {
    // Action
    QuickAssist = 'quick-assist',
    BasicAttack = 'basic-attack',
    DashAttack = 'dash-attack',
    DodgeCounter = 'dodge-counter',
    ExSpecialAttack = 'ex-special-attack',
    ChainAttack = 'chain-attack',
    DefensiveAssist = 'defensive-assist',
    EvasiveAssist = 'evasive-assist',
    Aftershock = 'aftershock',
    Ultimate = 'ultimate',
    // Anomaly
    Freeze = 'freeze',
    Shatter = 'shatter',
    Burning = 'burning',
    Assault = 'assault',
    Shocked = 'shocked',
}