// src/rpg-core/types/combat-stat.ts
export enum CombatStat {
    // Base stats
    Attack = 'attack',
    Defense = 'defense',
    Health = 'health',
    EnergyRegen = 'energy-regen',
    CriticalRate = 'critical-rate',
    AnomalyMastery = 'anomaly-mastery',
    AnomalyProficiency = 'anomaly-proficiency',
    PenRatio = 'pen-ratio',
    Daze = 'daze',
    Impact = 'impact',
    AnomalyBuildupResistance = 'anomaly-buildup-resistance',
    // Damages
    Damage = 'damage',
    Shield = 'shield',
    FireDamage = 'fire-damage',
    ElectricDamage = 'electric-damage',
    EtherDamage = 'ether-damage',
    PhysicalDamage = 'physical-damage',
    IceDamage = 'ice-damage',
    CorruptionAdditionalDamage = 'corruption-additional-damage',
    CriticalDamage = 'critical-damage',
    BasicAttackDamage = 'basic-attack-damage',
    DashAttackDamage = 'dash-attack-damage',
    AftershockDamage = 'aftershock-damage',
    ExSpecialAttackDamage = 'ex-special-attack-damage',
    AssistAttackDamage = 'assist-attack-damage',
    UltimateDamage = 'ultimate-damage',
    SheerDamage = 'sheer-damage',
}